Refactor/cleanup, need to fix broken crossover in GA and complete implementation for #5
